+++ user-service/src/main/java/com/ementor/profile/service/service/StudentProfileService.java
@@ -7,34 +7,25 @@ import com.ementor.profile.service.core.entity.pagination.filter.FilterOption;
 import com.ementor.profile.service.core.entity.pagination.filter.FilterOptionUtils;
 import com.ementor.profile.service.core.entity.pagination.filter.FilterType;
 import com.ementor.profile.service.core.exceptions.EmentorApiError;
-import com.ementor.profile.service.core.redis.entity.StoredRedisToken;
-import com.ementor.profile.service.core.redis.services.StoredRedisTokenService;
 import com.ementor.profile.service.core.service.SecurityService;
 import com.ementor.profile.service.dto.ProfilePrerequrireDTO;
 import com.ementor.profile.service.dto.StudentProfileDTO;
-import com.ementor.profile.service.entity.ProfilePicture;
 import com.ementor.profile.service.entity.StudentProfile;
 import com.ementor.profile.service.entity.StudentProfileView;
 import com.ementor.profile.service.entity.User;
 import com.ementor.profile.service.enums.RoleEnum;
 import com.ementor.profile.service.repo.StudentProfilesRepo;
 import com.ementor.profile.service.repo.StudentProfilesViewRepo;
-import com.ementor.profile.service.utils.ConstantUtils;
 import jakarta.persistence.EntityManager;
-import java.io.IOException;
-import java.util.List;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.core.io.ByteArrayResource;
 import org.springframework.data.domain.Page;
-import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.client.RestTemplate;
+
+import java.util.List;
+import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
@@ -55,16 +46,16 @@ public class StudentProfileService {
 
 	private final LocationService locationService;
 
-	private final StoredRedisTokenService storedRedisTokenService;
+	private final UserServiceRestService userServiceRest;
 
 	private final StudentProfilesRepo studentProfilesRepo;
 
 	private final StudentProfilesViewRepo studentProfilesViewRepo;
 
-	// TODO create user view and make paginated.
 	public PaginatedResponse<StudentProfileView> getPaginated(PaginatedRequest request) {
 		securityService.hasAnyRole(RoleEnum.ADMIN, RoleEnum.PROFESSOR);
 		User user = securityService.getCurrentUser();
+		String filterCriteria = "professorId";
 
 		log.info("[USER-ID:{}] Getting student profiles list - page {}", user.getUserId(), request.getPage());
 
@@ -72,13 +63,13 @@ public class StudentProfileService {
 			.genericSpecification(request.bind(StudentProfileView.class), false, StudentProfileView.class);
 		if (user.getRole()
 			.equals(RoleEnum.PROFESSOR)) {
-			PaginationUtils.addFilterCriteria(request, "professorId", user.getUserId());
+			PaginationUtils.addFilterCriteria(request, filterCriteria, user.getUserId());
 		}
 		Page<StudentProfileView> findAll = studentProfilesViewRepo.findAll(spec, ServiceUtils.convertToPageRequest(request));
 		PaginatedResponse<StudentProfileView> paginatedResponse = new PaginatedResponse<>();
 		if (user.getRole()
 			.equals(RoleEnum.PROFESSOR)) {
-			PaginationUtils.removeFilterCriteria(request, "professorId");
+			PaginationUtils.removeFilterCriteria(request, filterCriteria);
 		}
 		paginatedResponse.setData(findAll.getContent()
 			.stream()
@@ -122,6 +113,7 @@ public class StudentProfileService {
 
 		StudentProfile studentProfile = getStudentProfileById(studentProfileId);
 		StudentProfileDTO studentProfileDTO = buildStudentProfileDto(studentProfile);
+		studentProfileDTO.setUserId(studentProfile.getUserId());
 		studentProfileDTO.setId(studentProfile.getId());
 
 		log.info("[USER-ID: {}] Got student profile with id {}.", currentUserId, studentProfileId);
@@ -156,7 +148,7 @@ public class StudentProfileService {
 
 		studentProfilesRepo.save(studentProfile);
 
-		sendProfilePictureThumbnail();
+		userServiceRest.sendProfilePictureThumbnail();
 
 		log.info("[USER-ID: {}] Created  student profile.", currentUserId);
 	}
@@ -175,82 +167,7 @@ public class StudentProfileService {
 			.build();
 	}
 
-	private void sendProfilePictureThumbnail() {
-		User user = securityService.getCurrentUser();
-		ProfilePicture profilePicture = profilePictureService.getImageByUserId(user.getUserId());
-		byte[] reducedFilePicture = null;
-		try {
-			reducedFilePicture = profilePictureService.getPreparedIamge(profilePicture.getFileData(), 100, 100);
-		} catch (IOException e) {
-			throw new EmentorApiError("Could not get bytes of data. Either cannot save or cannot compress.", 404);
-		}
-		StoredRedisToken storedRedisToken = storedRedisTokenService.getStoredRedisToken(user.getUsername())
-			.orElseThrow(() -> new EmentorApiError("Token not found. Check if it is last token or valid token."));
-		if (storedRedisToken.getToken() == null) {
-			log.error("StoredRedisToken error: Token not found for user [{}]", user.getUserId());
-			return;
-		}
 
-		Boolean hasUploadedThumbnail = sendPostWithFile(reducedFilePicture, storedRedisToken, profilePicture.getFileName());
-		Boolean hasProfileCompleted = studentProfilesRepo.findByUserId(user.getUserId())
-			.isPresent();
-
-		if (hasUploadedThumbnail && hasProfileCompleted) {
-			HttpHeaders headers = new HttpHeaders();
-			headers.setBearerAuth(storedRedisToken.getToken());
-
-			HttpEntity<?> requestEntity = new HttpEntity<>(headers);
-
-			RestTemplate restTemplate = new RestTemplate();
-			String serverUrl = ConstantUtils.USER_SERVICE_PROD_URL + "/profile-data/profile-completed/true";
-
-			ResponseEntity<Boolean> response = restTemplate.exchange(serverUrl, HttpMethod.GET, requestEntity,
-					Boolean.class);
-
-			Boolean responseBody = response.getBody();
-			log.info("User {} completed the successful: {}", user.getEmail(), responseBody);
-		}
-	}
-
-	public Boolean sendPostWithFile(byte[] reducedFilePicture,
-			StoredRedisToken storedRedisToken,
-			String originalFilename) {
-		HttpHeaders headers = new HttpHeaders();
-		headers.setContentType(MediaType.MULTIPART_FORM_DATA);
-		headers.setBearerAuth(storedRedisToken.getToken());
-
-		ByteArrayResource resource = new ByteArrayResource(reducedFilePicture) {
-			@Override
-			public String getFilename() {
-				return "thumb_" + originalFilename;
-			}
-		};
-
-		MultiValueMap<String, Object> body = new LinkedMultiValueMap<>();
-		body.add("file", resource);
-
-		HttpEntity<MultiValueMap<String, Object>> requestEntity = new HttpEntity<>(body, headers);
-
-		RestTemplate restTemplate = new RestTemplate();
-		String serverUrl = ConstantUtils.USER_SERVICE_PROD_URL + "/profile-data/upload";
-		ResponseEntity<String> response = restTemplate.postForEntity(serverUrl, requestEntity, String.class);
-
-		if (response.getStatusCode() == HttpStatus.OK) {
-			String isUploaded = response.getBody();
-			try {
-				isUploaded = isUploaded.replace("\"", "");
-				UUID uuid = UUID.fromString(isUploaded);
-				log.info("Post success, image uploaded with id: {}", uuid);
-				return true;
-			} catch (IllegalArgumentException e) {
-				throw new EmentorApiError("The upload was not successful, UUID not found");
-			}
-		} else {
-			log.error(response.getStatusCode()
-				.toString());
-			return false;
-		}
-	}
 
 	public void updateStudentProfile(StudentProfileDTO dto) {
 		securityService.hasAnyRole(RoleEnum.STUDENT);
